Update Helm tiller version
This isn't the most recent version available (2.14.3). This version has a bug handling null Chart values https://github.com/helm/helm/issues/5184 which is a regression. We use this feature in our Chart, so this is problematic. Fix is already merged into master so it'll be available in the next version